Simple Mobile Phone App
Created a program that implements a simple mobile phone with the ability to store, modify, remove and query contacts by developing a functional menu of options that are available in order to control the application. Programmed the application to check when adding or updating contacts, if the contact already exists in the application.